refactor: detial.ts 코드 모듈화 및 recentPosts 추가
