+++ src/pages/detail-page/modules/profile.ts
@@ -1,0 +1,6 @@
+export function getProfileImage(src?: string) {
+  if (!src || src.trim() === '') {
+    return './../../../icons/logo.svg';
+  }
+  return src;
+}
