+++ src/pages/write-page/write.css
@@ -146,11 +146,11 @@ main textarea:placeholder-shown {
   background: rgba(0, 0, 0, 0.5);
   border: none;
   color: white;
-  width: 15px;
-  height: 15px;
+  width: 20px;
+  height: 20px;
   border-radius: 50%;
   cursor: pointer;
-  font-size: 10px;
+  font-size: 14px;
   display: flex;
   align-items: center;
   justify-content: center;
